Modified all numeric types models columns to float type. Created and implemented a new endpoint to query if a specific product is available at a certain store

+++ app/models/product.py
@@ -7,7 +7,7 @@ class Product(db.Model):
     __tablename__ = 'product'
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100), nullable=False, unique=True)
-    price = db.Column(db.Numeric(8, 2))
+    price = db.Column(db.Float)
     description = db.Column(db.String(255))
     categories = db.relationship(
         'Category', secondary=category_product, back_populates='products')
